fixed some bugs that prevented find best projection from working properly

+++ utils.h
@@ -8,7 +8,7 @@
 #ifndef MIN
 	#define MIN(a,b) a < b ? a : b
 #endif
-void computeIsolationDistance(float *data, float *means, unsigned int nrows, unsigned int ncols, unsigned int* cids, unsigned int nclusters, unsigned int *npoints, unsigned int* dims,float *minIsoDist);
+void computeIsolationDistance(float *data, float *means, unsigned int nrows, unsigned int ncols, unsigned int* cids, unsigned int nclusters, unsigned int *npoints, unsigned int* dims,double *minIsoDist);
 unsigned int *histogram_sorted(float *data, int datal, float*bins, int binsl,unsigned int *counts);
 int matrix_inverse(double *A,int N,double *det,int *sign);
 double chi2_cdf(double x, int df);
